catsnifferv2/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-jinja2==1.6
aiosignal==1.4.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
frozenlist==1.8.0
idna==3.11
intelhex==2.3.0
Jinja2==3.1.6
linkify-it-py==2.0.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdit-py-plugins==0.5.0
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
platformdirs==4.5.0
propcache==0.4.1
Pygments==2.19.2
pyserial==3.5
pyserial-asyncio==0.6
pyusb>=1.2.1
python-magic==0.4.27
requests==2.32.5
rich==14.2.0
shellingham==1.5.4
textual==6.5.0
textual-dev==1.8.0
textual-serve==1.1.3
typer==0.20.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
urllib3==2.5.0
yarl==1.22.0
